fix: conflicting replacers don't match on retry #8
This modification introduces new handling logic for
when a website has multiple replacers with different
results set for the same regex. This allows users to
handle websites that tend to return different urls
for the same resource, therefore requiring a retry
with a different selection.

Think of the scenario when a thumbnail is at
/thumbs/imageId.jpg and the full resolution version
could be available at either /images/imageId.jpg or
/images/imageId.png.

Closes #8

Signed-off-by: Nabil Tharwat <nabil.tharwat@outlook.com>